requirements.txt
flask
flask-cors
requests
python-dotenv
sqlalchemy
psycopg2-binary
langdetect
